Re-structured entire Duke class with moving code out of main to abstracted layers of functions
