+++ data/duke.txt
@@ -1,0 +1,2 @@
+List
+[T] ✘ apple
